rename : home 이외의 곳에서도 사용될 클래스 이동
