--- app/src/main/java/com/example/potatoservice/ui/home/AdapterCallback.kt
@@ -1,5 +1,0 @@
-package com.example.potatoservice.ui.home
-
-interface AdapterCallback {
-	fun onClicked(service: Service)
-}
